fix: triggering deploy action on new release
Context: It appears that on.push.tags doesn't get triggered on a new release. Using the release trigger is another option, but there we cannot filter by tag. Thus, the filtering needs to happen on the action itself, which this commit is about.
This means that with a new release of our creates, all trigger a github action, but three of them are skipped, while only the demo server succeeds.
In addition, I also would like to be able to trigger this action manually, so I added this as well - which complicated the logic a bit more.