fix: simplify github deploy action
Context: The last job was still skipped, even though the fly.io deployment succeeded. Most likely due to the DAG, since the last job depended on deploy, which depended on a skipped job. Instead of including yet another horrible if statement for the last job, I decided to just add it as a step to the deploy job.