feat: rewrite of core data model and architecture
This squashed commit represents an effort that started as a simple
experiment with 2 aims: (1) better separate "responsibilities"
throughout the system architecture, and (2) simplify and better align
the core data model to the flow trading whitepaper.

These aims were achieved, but the core data model, being as deeply
baked as it was, required significant rewrites of virtually every
component of this project. Functionally, things are mostly the same; the
API routes look a little different, but the Big Ideas remain.

A few features have (temporarily) been removed: most notably, there is no
longer support for specifying a "maximum/minimum trade bound" for a
portfolio, as all results are strictly interpreted as rates (no more
scaling to absolute amounts), and there is not presently any support for
a "bidder stream" SSE endpoint that streams auction outcomes to
listening parties. This latter endpoint will be rebuilt soon, but the
former constraint capability will likely not make a comeback.

A nice consequence of this rearchitecting is the increased ease of
testing, which has led to (and will continue to lead to more)
drastically improved test coverage.