feat: dedicated types for demand and product groups
This commit was surprisingly large for what it does:
Recognizing that Map is an implementation detail, it instead
defines semantic types DemandGroup and ProductGroup.

This instantly makes the OpenAPI schema nice, as well
as opens the door to future performance optimizations.

While updating the SQLx queries, the batch generation query
was also simplified somewhat.